backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "google-genai>=1.2.0",
    "httpx==0.27.2",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "python-multipart==0.0.9",
    "uvicorn[standard]==0.30.6",
    "yt-dlp",
    "pymongo",
    "bcrypt",
    "python-jose[cryptography]",
]
